working delay detection and udp ping

+++ core/dojo/DojoLobby.hpp
@@ -20,8 +20,8 @@ public:
 	std::string ConstructMsg();
 
 	std::map<std::string, std::string> active_beacons;
-	std::map<std::string, int> active_beacon_ping;
-	std::map<std::string, long> last_seen;
+	std::map<std::string, uint64_t> active_beacon_ping;
+	std::map<std::string, uint64_t> last_seen;
 };
 
 constexpr short multicast_port = 7776;
